requirements.txt
Django==1.7
django-tastypie==0.12.1
